// nodes/Smartling/common/files/file-types.ts
export enum FileType {
    ANDROID = "android",
    IOS = "ios",
    GETTEXT = "gettext",
    HTML = "html",
    JAVA_PROPERTIES = "java_properties",
    XLIFF = "xliff",
    XML = "xml",
    JSON = "json",
    DOCX = "docx",
    PPTX = "pptx",
    XLSX = "xlsx",
    IDML = "idml",
    INDD = "indd",
    QT = "qt",
    RESX = "resx",
    PLAIN_TEXT = "plain_text",
    CSV = "csv",
    SRT = "srt",
    STRINGSDICT = "stringsdict",
    XLS = "xls",
    DOC = "doc",
    PPT = "ppt",
    PRES = "pres",
    MADCAP = "madcap",
    YAML = "yaml",
    MARKDOWN = "markdown",
    PHP_RESOURCE = "php_resource",
    FREEMARKER = "freemarker",
    TMX = "tmx",
    XLIFF_CAT = "xliff_cat",
    NONE = "none",
    DITA = "dita",
    DITA_ZIP = "dita_zip",
    VTT = "vtt",
    PDF = "pdf",
    RTF = "rtf",
    FLARE = "flare",
    FLUENT = "fluent",
    XLSX_TEMPLATE = "xlsx_template",
    SVG = "svg",
    DOCM = "docm",
    XLIFF2 = "xliff2",
    ARB = "arb",
    XCSTRINGS = "xcstrings",
    VSDX = "vsdx",
    VSDM = "vsdm"
}

export const fileDescriptionMap: Record<FileType, string> = {
    [FileType.ANDROID]: "Android XML",
    [FileType.IOS]: "iOS Strings",
    [FileType.GETTEXT]: "Gettext",
    [FileType.HTML]: "HTML",
    [FileType.JAVA_PROPERTIES]: "Java Properties",
    [FileType.XLIFF]: "XLIFF",
    [FileType.XML]: "XML",
    [FileType.JSON]: "JSON",
    [FileType.DOCX]: "Microsoft Word",
    [FileType.PPTX]: "Microsoft PowerPoint",
    [FileType.XLSX]: "Microsoft Excel",
    [FileType.IDML]: "Adobe InDesign File (IDML)",
    [FileType.INDD]: "Adobe InDesign Documents (INDD)",
    [FileType.QT]: "QT Linguist",
    [FileType.RESX]: "RESX",
    [FileType.PLAIN_TEXT]: "Plain Text (TXT)",
    [FileType.CSV]: "CSV",
    [FileType.SRT]: "SubRip Subtitle (SRT)",
    [FileType.STRINGSDICT]: "iOS Stringsdict",
    [FileType.XLS]: "Microsoft Excel Document",
    [FileType.DOC]: "Microsoft Word Document",
    [FileType.PPT]: "Microsoft Powerpoint Presentation",
    [FileType.PRES]: "Presentation File",
    [FileType.MADCAP]: "MadCap Lingo ZIP Package",
    [FileType.YAML]: "YAML",
    [FileType.MARKDOWN]: "Markdown",
    [FileType.PHP_RESOURCE]: "PHP resources",
    [FileType.FREEMARKER]: "FreeMarker template",
    [FileType.TMX]: "TMX translation memory file format",
    [FileType.XLIFF_CAT]: "XLIFF for offline CAT integration",
    [FileType.NONE]: "",
    [FileType.DITA]: "DITA",
    [FileType.DITA_ZIP]: "DITA zip file format",
    [FileType.VTT]: "WebVTT",
    [FileType.PDF]: "PDF*",
    [FileType.RTF]: "Rich Text Format (RTF)",
    [FileType.FLARE]: "MadCap Flare ZIP Package",
    [FileType.FLUENT]: "Fluent Localization",
    [FileType.XLSX_TEMPLATE]: "Microsoft Excel Template",
    [FileType.SVG]: "Scalable Vector Graphics (SVG)",
    [FileType.DOCM]: "Word Macro-Enabled Document",
    [FileType.XLIFF2]: "XLIFF 2.0",
    [FileType.ARB]: "Flutter Application Resource Bundle",
    [FileType.XCSTRINGS]: "iOS String Catalog",
    [FileType.VSDX]: "Microsoft Visio",
    [FileType.VSDM]: "Microsoft Visio Drawing File Format"
};

export const toFileTypeChoices = (types: FileType[]) => types.map(type => ({
    sample: type,
    value: type,
    label: fileDescriptionMap[type]
})).sort((a, b) => b.label.localeCompare(a.label));